epa2data control null value for selector

+++ dbmodel/ws/fct/ws_gw_fct_epa2data.sql
@@ -54,8 +54,10 @@ BEGIN
   FROM v_rpt_node a 
   JOIN connec ON node_id = connec_id WHERE result_id=v_result_id;
 
-	DELETE FROM selector_rpt_main WHERE cur_user=current_user;
-	INSERT INTO selector_rpt_main(result_id, cur_user) VALUES (v_current_selector, current_user);
+  IF v_current_selector IS NOT NULL THEN
+		DELETE FROM selector_rpt_main WHERE cur_user=current_user;
+		INSERT INTO selector_rpt_main(result_id, cur_user) VALUES (v_current_selector, current_user);
+	END IF;
 
   	--  Return
 	RETURN gw_fct_json_create_return(('{"status":"Accepted", "message":{"level":1, "text":"Analysis done successfully"}, "version":"'||v_version||'"'||
@@ -64,9 +66,9 @@ BEGIN
 			'}}'||
 	    '}')::json, 3180, null, null, null);
 
-	EXCEPTION WHEN OTHERS THEN
-	GET STACKED DIAGNOSTICS v_error_context = PG_EXCEPTION_CONTEXT;
-	RETURN ('{"status":"Failed","NOSQLERR":' || to_json(SQLERRM) || ',"SQLSTATE":' || to_json(SQLSTATE) ||',"SQLCONTEXT":' || to_json(v_error_context) || '}')::json;
+--	EXCEPTION WHEN OTHERS THEN
+--	GET STACKED DIAGNOSTICS v_error_context = PG_EXCEPTION_CONTEXT;
+--	RETURN ('{"status":"Failed","NOSQLERR":' || to_json(SQLERRM) || ',"SQLSTATE":' || to_json(SQLSTATE) ||',"SQLCONTEXT":' || to_json(v_error_context) || '}')::json;
 		
 END;
 $BODY$
